feat(db): move sqlite_migrator into its own module
